docs/requirements.txt
librosa
numpy
scipy
torch
fastapi
uvicorn
python-multipart
requests
pymongo
transformers
mediapipe
opencv-python-headless
nltk
sentence-transformers
google-generativeai
moviepy==1.0.3
streamlit
python-dotenv
gunicorn
openai-whisper
torchvision
facenet-pytorch
gradio